requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
auth0-python==4.10.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Bottleneck==1.5.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.2.1
cohere==5.15.0
coloredlogs==15.0.1
crewai==0.130.0
crewai-tools==0.47.1
cryptography==45.0.4
dataclasses-json==0.6.7
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
durationpy==0.10
embedchain==0.1.128
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.13
fastavro==1.11.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.97.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.20.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gptcache==0.1.44
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.4
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.8.3
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
json5==0.12.0
json_repair==0.46.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
lancedb==0.23.0
langchain==0.3.25
langchain-cohere==0.3.5
langchain-community==0.3.25
langchain-core==0.3.65
langchain-experimental==0.3.4
langchain-google-vertexai==2.0.25
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langsmith==0.3.45
litellm==1.72.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.109
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.5.0
mypy_extensions==1.1.0
networkx==3.5
nodeenv==1.9.1
numexpr==2.11.0
numpy==2.3.0
oauthlib==3.3.0
onnxruntime==1.22.0
openai==1.88.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.3.0
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.2.1
portalocker==2.10.1
posthog==3.25.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pypdf==5.6.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.402
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytube==15.0.0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
qdrant-client==1.14.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
schema==0.7.7
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.20.3
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uv==0.7.13
uvicorn==0.34.3
uvloop==0.21.0
validators==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0

crewai>=0.28.8
crewai[tools]>=0.1.2
langchain-google-genai==1.0.2
schedule==1.2.1
google-api-python-client==2.126.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
moviepy==1.0.3
replicate==0.25.1
elevenlabs==1.2.0
google-cloud-texttospeech==2.15.0
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
diffusers[torch]==0.27.2
transformers==4.40.1
accelerate==0.29.3
dlib-bin==19.24.0
face-alignment==1.4.1
opencv-python==4.9.0.80
